refactor(components): improve code-splitting
