+++ src/components/Introduction/Introduction.module.scss
@@ -7,7 +7,8 @@
 }
 
 .title {
-  font-size: 2rem;
+  font-size: 8rem;
+  line-height: 1.2;
   font-weight: 400;
   text-align: center;
   margin: 0 auto;
